allow the cfgfile code to try multiple configuration files before to give up.
The configuration loading code must now be supplied with a NULL-terminated
array of directory to scan when asked to load a config file.

The configuration loading code attempt to traverse such array in order and
stops at the first file found.

the tc_config_set_dir now sets a common last-resort directory.